Update username regex to support only lowercase
